autoconf: Upgrade to 2.72c
2.72c is a prerelease version of autoconf 2.73. It contains largefile and y2038 64 bit
time_t improvements for 32 bit architectures.

Rather than work on the older codebase, this brings us to work with the recent
autoconf upstream with the 64 bit changes. It is unclear when upstream will release
2.73 but it is easier for us to be aligned now we've done the bulk of the work needed
to update.

Upstream added several patches which fixed several common failures OE builds ran
into (backported in the next commit). In general testing has otherwise been good for us.

There is an unfortunate gnulib largefile.m4 bug. This change patches various software
to workaround it, next time they update new versions of the gnulib code will be pulled
in which address the issue with the official fix.

There are also a couple of ordering related fixes for apr and libarchive.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>