lib: Import packagedata oe module by default
Variable dependencies for functions in the oe.packagedata module were
missing as it was not present in BBIMPORTS. Add it as the fact it was
missing is likely just historical oversight from base.bbclass history
and the dependencies are useful.

Add an exclusion to bitbake.conf to ensure BB_NUMBER_THREADS doesn't
change task checksums.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>