linux-yocto: introduce 6.5 reference kernel recipes
The v6.4 kernel series has been EOLd upstream, so it isn't a
suitable "latest" reference kernel for the release.

The -dev kernel has previously tested 6.5, so we know that
core sanity is acceptable.

We introduce the 6.5 kernel recipes to go along with a -dev
of 6.6 and LTS of v6.1

Testing against all arches and refernce BSPs has passed.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>